Move backend database tests
