+++ packages/backend/src/peripherals/database/PageMappingRepository.test.ts
@@ -1,12 +1,12 @@
 import { Hash256 } from '@explorer/types'
 import { expect } from 'earljs'
 
+import { setupDatabaseTestSuite } from '../../test/database'
+import { Logger } from '../../tools/Logger'
 import {
   PageMappingRecord,
   PageMappingRepository,
-} from '../../../src/peripherals/database/PageMappingRepository'
-import { Logger } from '../../../src/tools/Logger'
-import { setupDatabaseTestSuite } from './shared/setup'
+} from './PageMappingRepository'
 
 describe(PageMappingRepository.name, () => {
   const { database } = setupDatabaseTestSuite()
