+++ packages/backend/src/peripherals/database/PageRepository.test.ts
@@ -1,16 +1,13 @@
 import { Hash256 } from '@explorer/types'
 import { expect } from 'earljs'
 
+import { setupDatabaseTestSuite } from '../../test/database'
+import { Logger } from '../../tools/Logger'
 import {
   PageMappingRecord,
   PageMappingRepository,
-} from '../../../src/peripherals/database/PageMappingRepository'
-import {
-  PageRecord,
-  PageRepository,
-} from '../../../src/peripherals/database/PageRepository'
-import { Logger } from '../../../src/tools/Logger'
-import { setupDatabaseTestSuite } from './shared/setup'
+} from './PageMappingRepository'
+import { PageRecord, PageRepository } from './PageRepository'
 
 describe(PageRepository.name, () => {
   const { database } = setupDatabaseTestSuite()
