+++ packages/backend/src/peripherals/database/SyncStatusRepository.test.ts
@@ -1,9 +1,9 @@
 import { expect } from 'earljs'
 
-import { KeyValueStore } from '../../../src/peripherals/database/KeyValueStore'
-import { SyncStatusRepository } from '../../../src/peripherals/database/SyncStatusRepository'
-import { mock } from '../../../src/test/mock'
-import { Logger } from '../../../src/tools/Logger'
+import { mock } from '../../test/mock'
+import { Logger } from '../../tools/Logger'
+import { KeyValueStore } from './KeyValueStore'
+import { SyncStatusRepository } from './SyncStatusRepository'
 
 describe(SyncStatusRepository.name, () => {
   it('writes value to store', async () => {
